ros_code/freicar_object_detection/requirements.txt
numpy==1.16.0
opencv-python
tqdm
pyyaml
webcolors
matplotlib
torch
torchvision
webcolors
visdom